datavault-frontend/requirements.txt
streamlit==1.31.1
pandas==2.2.0
kaggle==1.6.6
requests==2.31.0
python-dotenv==1.0.1
supabase==1.0.3
extra-streamlit-components==0.1.60
plotly==5.18.0